Keep VkViewport out of the public API: EffectiveViewport returns it instead
Follow-up to the Viewport overload: the ViewportRect -> VkViewport
conversion operator put a (forward-declared) Vulkan type into the
API-neutral public header. Inverted instead -- the impl works in the
renderer's own form, the public API never learns about Vulkan:

- Impl::EffectiveViewport now returns VkViewport directly (0..1 depth,
  the recording-side convention); RenderPasses/RenderSetup consume its
  float members without casts.
- RenderContext::GetViewport maps it back to the API-neutral
  ViewportRect (whole-pixel values, so the narrowing is exact).
- ViewportRect is a plain aggregate again; Zahlen/Render.hpp compiles
  with no Vulkan header in sight.

+++ src/engine/graphics/RenderPasses.cpp
@@ -269,8 +269,8 @@ struct GpuCullingPolicyPass1 {
 
         const auto sceneVp = ctx.EffectiveViewport();
         pc.viewProj         = ctx.unjittered_view_proj;
-        pc.hizScreenSize[0] = static_cast<float>(sceneVp.width);
-        pc.hizScreenSize[1] = static_cast<float>(sceneVp.height);
+        pc.hizScreenSize[0] = sceneVp.width;
+        pc.hizScreenSize[1] = sceneVp.height;
         const uint32_t hizMips = std::min(ctx.graphResources.hizMap.mipLevels, kMaxGeneratedHiZMips);
         pc.maxHiZMipLevel   = hizMips > 0 ? hizMips - 1 : 0;
         pc.drawCount        = drawCount;
@@ -352,7 +352,7 @@ struct GpuCullingPolicyPass2 {
         const auto sceneVp2 = ctx.EffectiveViewport();
         RenderContext::Impl::CullingConstants pc {
             .viewProj       = ctx.unjittered_view_proj,
-            .hizScreenSize  = {static_cast<float>(sceneVp2.width), static_cast<float>(sceneVp2.height)},
+            .hizScreenSize  = {sceneVp2.width, sceneVp2.height},
             .maxHiZMipLevel = hizMips2 > 0 ? hizMips2 - 1 : 0,
             .drawCount      = drawCount,
             .passIndex      = 1,
